Remove nltk from standard graph extraction
